+++ packages/styles/src/app/vanilla-extract/VanillaExtractSkeleton.tsx
@@ -1,12 +1,9 @@
-import { useTheme } from '../../hooks/useTheme';
 import { vanillaExtractSkeletonItem, vanillaExtractSkeletonWrapper } from './skeleton.css';
 
 export const VanillaExtractSkeleton = () => {
-  const { theme } = useTheme();
-
   return (
     <div className={vanillaExtractSkeletonWrapper}>
-      <span className={vanillaExtractSkeletonItem[theme]}>Skeleton</span>
+      <span className={vanillaExtractSkeletonItem}>Skeleton</span>
     </div>
   );
 };